Modularized, Human Test, Mail, SVGs
Third time ive had to write this...
Modularized the site, now that I know more Php.

Implemented Mail for the contact form so any enquires made are sent to
my alternate Gmail account.

Added a Human test to the contact form in an attempt to prevent
automated submissions. Users need to identify an image next to the input
field in order to submit an enquiry.

Replaced the images on the welcome screen with SVGs, they now resize
well, are crisp and are significantly smaller in size. Still need to
figure out how to display the original pngs if the svgs can't be
displayed.

Removed the gallery temporarily for reconstruction. I wasn't satisfied
with the slider as it wasn't well implemented. I'm thinking ill make a
grid of small images that load larger images when clicked on and add
bars/buttons under the grid images linking to the live site.