fixed contacts, revamped gallery
fixed the contacts page as that was not displaying, it was an improperly
closed script tag.

completed revamping the gallery, it is no longer a slider and is now a
collection of smaller images that when clicked on download and overlay a
larger image, that is closed by clicking, if the title is clicked it
takes the user to a example site I have set up that the image
represents; thinking i need someway of hinting that they can do this.