Fixed contact form messaging
I wasn't receiving the message people were entering into the contact
form. the code was looking for a contact element in the post array when
it is in a message element. corrected that and corrected some incorrect
validation logic.

Also improved the mobile look of the contact section and increased the
input element's height.